update components folder struct, redactored GameCard
added games and game directory, decomposed price from GameCard into GamePrice
