+++ client/src/pages/GameName.tsx
@@ -3,6 +3,7 @@ import MainLayout from "../components/layouts/MainLayout";
 import {useNavigate, useParams} from "react-router-dom";
 import {useAppDispatch, useAppSelector} from "../hooks/redux";
 import {fetchOneGame} from "../store/actions/gameAction";
+import {GamePrice} from "../components/games/game/GamePrice";
 
 const GameName = () => {
     const {name} = useParams()
@@ -15,14 +16,22 @@ const GameName = () => {
         else dispatch(fetchOneGame(name))
     }, [])
 
-    if(loading === "failed"){
+    if (loading === "failed") {
         navigate("/error")
     }
 
     return (
         <MainLayout>
             <h1 className={"text-3xl"}>{game.name}</h1>
-            <img src={game.base64Image}/>
+            <div className={"flex"}>
+                <div className={"text-white w-3/4"}>
+                    <img src={game.base64Image}/>
+                    <div>{game.description}</div>
+                </div>
+                <div>
+                    <GamePrice price={game.price} discount={game.discount}/>
+                </div>
+            </div>
         </MainLayout>
     );
 };
