+++ client/src/pages/Games.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GamesList from "../components/GamesList";
+import GamesList from "../components/games/GamesList";
 import MainLayout from "../components/layouts/MainLayout";
 
 const Games = () => {
